\name{insertval2}
\alias{insertval2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Secondary function to code data 
}
\description{
Secondary function to code data in complete disjunctive form 
}
\usage{
insertval2(x, nmod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Data matrix in reduced coding (primitive coding)}
  \item{nmod}{
number of categories of each variable}
}
\details{
It helps to return a matrix from reduced coding in complete disjunctive coding}
%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
\references{
Lombardo R and Meulman JJ (2010) Journal of Classification, 27, 191-210. \cr
Beh EJ Lombardo R (2014) Correspondence Analysis, Theory, Practice and New Strategies. Wiley}
\author{
Rosaria Lombardo}
%\note{
%%  ~~further notes~~
%}
%% ~Make other sections like Warning with \section{Warning }{....} ~
%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.
%## The function is currently defined as}
%\keyword{ ~kwd1 }
\keyword{multivariate}% __ONLY ONE__ keyword per line
